app/chat-api/requirements.txt
# Install shared lib first (from repo root): pip install -e libs/code-shared
# Then: pip install -r app/chat-api/requirements.txt

### Core RAG stack (vector store and semantic cache live in chat-api)
docling
llama-index
weaviate-client
redis

### API and web
fastapi[standard]
uvicorn

### Embeddings / rerankers
rank-bm25
cohere

### Chat memory (Cassandra client)
cassandra-driver

### Configuration / validation
pydantic
pydantic-settings

### Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0
httpx>=0.24.0
faker>=18.0.0

# Install shared lib first (from repo root): pip install -e libs/code-shared
# Then: pip install -r app/chat-api/requirements.txt

### Core RAG stack (vector store and semantic cache live in chat-api)
docling
llama-index
weaviate-client
redis

### API and web
fastapi[standard]
uvicorn

### Embeddings / rerankers
rank-bm25
cohere

### Persistence & auth
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

### Configuration / validation
pydantic
pydantic-settings

### Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0
httpx>=0.24.0
faker>=18.0.0